% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/emoji_get.R
\name{emoji_get}
\alias{emoji_get}
\title{Get an emoji image from its codepoint}
\usage{
emoji_get(input, size = 72)
}
\arguments{
\item{input}{either a codepoint (without \code{\\U} etc), or the result from \code{\link{emoji_search}}}

\item{size}{the size of the resulting image (72, 36, 16)}
}
\value{
list of images
}
\description{
Get an emoji image from its codepoint
}
\note{
Adapted from \code{rphylopic} code by Scott Chamberlain
}
\author{
David L Miller
}

