% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji_search.R
\name{emoji_search}
\alias{emoji_search}
\title{Search for emoji}
\usage{
emoji_search(search)
}
\description{
Do a simple search through emoji keywords and names for a given term.
}
\note{
Data are taken from \url{https://github.com/muan/emojilib}
}
\author{
David L Miller
}
