% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_emoji.R
\name{geom_emoji}
\alias{geom_emoji}
\title{Emoji geom}
\usage{
geom_emoji(
  mapping = NULL,
  data = NULL,
  stat = "emoji",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
Use emoji as you would \code{\link{geom_point}}.
}
\examples{
\dontrun{
library(ggplot2)
library(emoGG)

# I want a tulip to plot for the iris example data
emoji_search("tulip")

# plot the data, but with emojis
ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
  geom_emoji(emoji="1f337")

}
}
\author{
David L Miller
}
