% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_emoji.R
\name{add_emoji}
\alias{add_emoji}
\title{Input an emoji and create a ggplot2 layer to add to an existing plot}
\usage{
add_emoji(img, x = NULL, y = NULL, ysize = NULL)
}
\arguments{
\item{img}{A png object, e.g, from using \code{emoji_get}.}

\item{x}{x value of the emoji center. Ignored if y and ysize are not specified.}

\item{y}{y value of the emoji center. Ignored if x and ysize are not specified.}

\item{ysize}{Height of the emoji. The width is determined by the aspect ratio of the original image. Ignored if x and y are not specified.}
}
\description{
Input an emoji and create a ggplot2 layer to add to an existing plot
}
\details{
Use parameters \code{x}, \code{y}, and \code{ysize} to place the emoji at a specified position on the plot. If all three of these parameters are unspecified, then the emoji will be plotted to the full height and width of the plot.
}
\note{
Based on \code{add_phylopic} from \code{rphylopic} by Scott Chamberlain.
}
\examples{
\dontrun{
library(ggplot2)
library(emoGG)

# Put a emoji behind a plot
flower <- emoji_get("1f337" )[[1]]
qplot(x=Sepal.Length, y=Sepal.Width, data=iris, geom="point") + add_emoji(flower)

# Put a silhouette anywhere
posx <- runif(50, 0, 10)
posy <- runif(50, 0, 10)
sizey <- runif(50, 0.4, 2)

cat <- emoji_get("1f697" )[[1]]
p <- ggplot(data.frame(x = posx, y = posy), aes(x, y)) +
            geom_point(color = rgb(0,0,0,0))
for (i in 1:50) {
  p <- p + add_emoji(cat, posx[i], posy[i], sizey[i])
}
}
}

